requirements.txt
telegraph
git+https://github.com/Hrishi2861/pyrofork-2.2.11-peer-fix.git
TgCrypto
aiohttp
hurry.filesize
python-dotenv
pixeldrain-reloaded
pytz
swibots